+++ test/tests/chai-loader.js
@@ -3,6 +3,6 @@ window.waitChai = new Promise(resolve => {
 	resolveChai = resolve;
 });
 
-import('https://cdn.jsdelivr.net/npm/chai@5.1.2/+esm').then(({ expect }) =>
-	resolveChai(expect)
+import('https://cdn.jsdelivr.net/npm/chai@5.1.2/+esm').then(chai =>
+	resolveChai(chai)
 );